refactor(handy-core): move history repo, inject event sink + retention config
