+++ crates/handy-core/src/audio/utils.rs
@@ -1,1 +1,26 @@
-//! Reserved: platform-independent audio utilities will live here.
+use anyhow::Result;
+use hound::{WavSpec, WavWriter};
+use log::debug;
+use std::path::Path;
+
+/// Save audio samples as a WAV file
+pub async fn save_wav_file<P: AsRef<Path>>(file_path: P, samples: &[f32]) -> Result<()> {
+    let spec = WavSpec {
+        channels: 1,
+        sample_rate: 16000,
+        bits_per_sample: 16,
+        sample_format: hound::SampleFormat::Int,
+    };
+
+    let mut writer = WavWriter::create(file_path.as_ref(), spec)?;
+
+    // Convert f32 samples to i16 for WAV
+    for sample in samples {
+        let sample_i16 = (sample * i16::MAX as f32) as i16;
+        writer.write_sample(sample_i16)?;
+    }
+
+    writer.finalize()?;
+    debug!("Saved WAV file: {:?}", file_path.as_ref());
+    Ok(())
+}
