+++ src-tauri/Cargo.toml
@@ -27,6 +27,7 @@ serde = { version = "1", features = ["derive"] }
 
 [dependencies]
 handy-core = { path = "../crates/handy-core" }
+handy-platform = { path = "../crates/handy-platform" }
 once_cell = "1"
 tauri = { version = "2.10.2", features = [
   "protocol-asset",
